# swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponseFormat": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"error": {
						"type": "string"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SortType": {
				"description": "Sort types for the leaderboard",
				"enum": [
					"byDucks",
					"byMoney",
					"byProduction",
					"byStorage",
					"byNbBuildings",
					"byNbStorage"
				],
				"type": "string"
			},
...
			"RegisterBody": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterAdminBody": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"playerAuth": {
				"type": "http",
				"scheme": "bearer"
			},
			"adminAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "DuckCorp API",
		"version": "1.0.0",
		"description": "The world's leading rubber duck manufacturer.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "CapyBlaze",
			"email": "capyblaze@gmail.com"
		}
	},
	"paths": {
...
		"/leaderboard": {
			"get": {
				"operationId": "GetLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Leaderboard by ducks retrieved successfully",
											"data": [
												{
													"name": "DuckMaster",
													"ducks": 420
												},
												{
													"name": "QuackFactory",
													"ducks": 300
												}
											],
											"timestamp": "2026-06-17T18:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get a paginated leaderboard sorted by ducks, money, production, storage capacity, building count, or storage count.",
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"playerAuth": []
					}
				],
				"parameters": [
					{
						"description": "The criteria to sort the leaderboard by (default: byDucks)",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortType"
						}
					},
					{
						"description": "The page number for pagination (default: 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"description": "The number of players to return per page (default: 10)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1,
							"maximum": 20
						}
					}
				]
			}
		},
		"/leaderboard/me": {
			"get": {
				"operationId": "GetMyLeaderboard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "Your leaderboard ranks retrieved successfully",
											"data": {
												"rank": 3,
												"ducksRank": 4,
												"moneyRank": 6,
												"productionRank": 3,
												"storageRank": 5,
												"nbBuildingsRank": 2,
												"nbStorageRank": 7
											},
											"timestamp": "2026-06-17T18:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Get the authenticated player's rank in every leaderboard category.",
				"tags": [
					"Leaderboard"
				],
				"security": [
					{
						"playerAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "User registered",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "User registered",
											"data": {
												"id": "cmz8n7r2g0000v9k4a1b2c3d4",
												"token": "player_eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
												"name": "DuckMaster",
												"lastActive": "2026-06-17T18:30:00.000Z",
												"createdAt": "2026-06-17T18:30:00.000Z",
												"updatedAt": "2026-06-17T18:30:00.000Z"
											},
											"timestamp": "2026-06-17T18:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing name",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Register a new player account.\nThe response includes the bearer token used to authenticate player routes.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterBody"
							}
						}
					}
				}
			}
		},
...
		"/admin/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"message": "User logged in",
											"data": {
												"token": "admin_eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9",
												"name": "admin"
											},
											"timestamp": "2026-06-17T18:30:00.000Z"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Missing credentials",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponseFormat"
								}
							}
						}
					}
				},
				"description": "Authenticate as an administrator and receive the bearer token required for admin routes.",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterAdminBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}